backend/requirements.txt
# FastAPI and ASGI server
fastapi
gunicorn
uvicorn[standard]
python-multipart

# Database
sqlalchemy
psycopg2-binary
alembic
motor
pymongo

# Authentication & Security
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1

# Environment & Configuration
pydantic
pydantic-settings
email-validator
python-dotenv

# Payment Gateway
razorpay

# File handling
pillow

# HTTP client (for testing)
httpx

# Development tools
pytest
pytest-asyncio
